Support 5.6.2 - no utf8, problems with PERL_DL_NONLAZY=1 and prefer_integer
utf8 tests are skipped because of missing API functions, but all cPanel tests work fine.
utf8 strings are stored and restored as raw bytes, they will loose the utf8 flag.

prefer_integer does not work with PERL_DL_NONLAZY=1. No idea which side-effect is causing this.

5.6 has a broken do. it returns not the value of the last expression,
but the subvalue in an array (t/data.pl hack)

5.6 has no PerlIO